requirements.txt
django>=1.5.1
wheel==0.24.0
# Additional requirements go here
django-scar
django-internationalflavor